Working simulation, making tweaks to api
