+++ _examples/simulate.go
